+++ lib/modelviews/constants.dart
@@ -1,9 +1,9 @@
 import "package:flutter/material.dart";
 
-const double roomWithTV = 4;
-const double roomWithSound = 3;
-const double roomWithDataShow = 4;
-const double roomWithWifi = 5;
+const int roomWithTV = 4;
+const int roomWithSound = 3;
+const int roomWithDataShow = 4;
+const int roomWithWifi = 5;
 double defaultpd = 16;
 const Color bgColor = Color(0xFF184848);
 const Color bgButton = Color(0xFF006060);
